InfraStructureLayerWithDtoInApplicationLayer

+++ src/Services/Ordering/Ordering.Domain/Models/OrderItem.cs
@@ -4,7 +4,7 @@ using System.ComponentModel.DataAnnotations;
 namespace Ordering.Domain.Models;
 public class OrderItem:Entity<OrderItemId>
 {
-    internal OrderItem(OrderId orderId, ProductId productId, int quantity, decimal price)
+    internal OrderItem(OrderId orderId, ProductId productId, int quantity, double price)
     {
         Id = OrderItemId.of(Guid.NewGuid());
         OrderId = orderId;
@@ -15,5 +15,5 @@ public class OrderItem:Entity<OrderItemId>
     public OrderId OrderId { get; private set; } = default!;
     public ProductId ProductId { get; private set; } = default!;
     public int Quantity { get; private set; } = default!;
-    public decimal Price { get; private set;} = default!;
+    public double Price { get; private set;} = default!;
 }
